# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CustomerPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Customer"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CustomerPage",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NewCustomer": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "maxLength": 200,
            "title": "Display Name",
            "type": "string"
          },
          "phone": {
            "anyOf": [
              {
                "maxLength": 40,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "required": [
          "display_name"
        ],
        "title": "NewCustomer",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Qarz Daftari",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/v1/shops/{shop_id}/customers/{customer_id}": {
      "get": {
        "operationId": "admin_support_customers_read_api_admin_v1_shops__shop_id__customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Admin Support Customers Read Api Admin V1 Shops  Shop Id  Customers  Customer Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin.Support.Customers.Read"
      }
    },
...
    "/api/v1/shops/{shop_id}/customers": {
      "get": {
        "operationId": "customers_list_api_v1_shops__shop_id__customers_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "active",
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 400,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.List"
      },
      "post": {
        "operationId": "customers_create_api_v1_shops__shop_id__customers_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewCustomer"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Customers Create Api V1 Shops  Shop Id  Customers Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.Create"
      }
    },
    "/api/v1/shops/{shop_id}/customers/{customer_id}": {
      "get": {
        "operationId": "customers_read_api_v1_shops__shop_id__customers__customer_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.Read"
      },
      "patch": {
        "operationId": "customers_update_api_v1_shops__shop_id__customers__customer_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Customers Update Api V1 Shops  Shop Id  Customers  Customer Id  Patch",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.Update"
      }
    },
    "/api/v1/shops/{shop_id}/customers/{customer_id}/archive": {
      "post": {
        "operationId": "customers_archive_api_v1_shops__shop_id__customers__customer_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Customers Archive Api V1 Shops  Shop Id  Customers  Customer Id  Archive Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.Archive"
      }
    },
    "/api/v1/shops/{shop_id}/customers/{customer_id}/entries": {
      "post": {
        "operationId": "ledger_entry_create_api_v1_shops__shop_id__customers__customer_id__entries_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEntry"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ledger Entry Create Api V1 Shops  Shop Id  Customers  Customer Id  Entries Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger.Entry.Create"
      }
    },
...
    "/api/v1/shops/{shop_id}/customers/{customer_id}/unarchive": {
      "post": {
        "operationId": "customers_unarchive_api_v1_shops__shop_id__customers__customer_id__unarchive_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "customer_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Customer Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Customers Unarchive Api V1 Shops  Shop Id  Customers  Customer Id  Unarchive Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Customers.Unarchive"
      }
    },
...
    "/api/v1/shops/{shop_id}/entries/{entry_id}/lines": {
      "post": {
        "operationId": "ledger_entry_lines_add_api_v1_shops__shop_id__entries__entry_id__lines_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewLines"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ledger Entry Lines Add Api V1 Shops  Shop Id  Entries  Entry Id  Lines Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger.Entry.Lines.Add"
      }
    },
    "/api/v1/shops/{shop_id}/entries/{entry_id}/promise": {
      "post": {
        "operationId": "ledger_entry_promise_change_api_v1_shops__shop_id__entries__entry_id__promise_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromiseChange"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ledger Entry Promise Change Api V1 Shops  Shop Id  Entries  Entry Id  Promise Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger.Entry.Promise.Change"
      }
    },
    "/api/v1/shops/{shop_id}/entries/{entry_id}/promise-choice": {
      "post": {
        "operationId": "ledger_entry_promise_choose_api_v1_shops__shop_id__entries__entry_id__promise_choice_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromiseChoice"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ledger Entry Promise Choose Api V1 Shops  Shop Id  Entries  Entry Id  Promise Choice Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger.Entry.Promise.Choose"
      }
    },
    "/api/v1/shops/{shop_id}/entries/{entry_id}/reversal": {
      "post": {
        "operationId": "ledger_entry_reverse_api_v1_shops__shop_id__entries__entry_id__reversal_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "entry_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Entry Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Ledger Entry Reverse Api V1 Shops  Shop Id  Entries  Entry Id  Reversal Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ledger.Entry.Reverse"
      }
    },
...
    "/api/v1/shops/{shop_id}/overview": {
      "get": {
        "operationId": "overview_read_api_v1_shops__shop_id__overview_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Overview"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Overview.Read"
      }
    },
...
  }
}